// client/src/typeScript/pam.ts
export interface PamMockData {
    id: number;
    gender: string;
    language: string;
    date_bl: Date;
    date_m3: Date;
    date_m6: Date;
    date_m9: Date;
    date_m12: Date;
    raw_bl: number;
    raw_m3: number;
    raw_m6: number;
    raw_m9: number;
    raw_m12: number;
    activation_bl: number;
    activation_m3: number;
    activation_m6: number;
    activation_m9: number;
    activation_m12: number;
    level_bl: number;
    level_m3: number;
    level_m6: number;
    level_m9: number;
    level_m12: number;
}

export interface PamResults {

}